slight store API and sources directory refactoring
