+++ SwiftyRedux/Sources/Command/Redux+Command.swift
@@ -4,15 +4,15 @@
 
 extension Store {
     @discardableResult
-    public func subscribe(on queue: DispatchQueue? = nil, _ command: Command<State>) -> Disposable {
-        return subscribe(on: queue, observer: command.execute)
+    public func subscribe(on queue: DispatchQueue? = nil, includingCurrentState: Bool = true, _ command: Command<State>) -> Disposable {
+        return subscribe(on: queue, includingCurrentState: includingCurrentState, observer: command.execute)
     }
 }
 
 extension Store where State: Equatable {
     @discardableResult
-    public func subscribe(on queue: DispatchQueue? = nil, skipRepeats: Bool = false, _ command: Command<State>) -> Disposable {
-        return subscribe(on: queue, skipRepeats: skipRepeats, observer: command.execute)
+    public func subscribeUnique(on queue: DispatchQueue? = nil, includingCurrentState: Bool = true, _ command: Command<State>) -> Disposable {
+        return subscribeUnique(on: queue, includingCurrentState: includingCurrentState, observer: command.execute)
     }
 }
 
@@ -22,6 +22,14 @@ extension Observable {
     @discardableResult
     public func subscribe(on observingQueue: DispatchQueue? = nil, _ command: Command<Value>) -> Disposable {
         return subscribe(on: observingQueue, observer: command.execute)
+    }
+}
+
+// MARK: - Observable Producer
+
+extension ObservableProducer {
+    public func start(on observingQueue: DispatchQueue? = nil, _ command: Command<Value>) -> Disposable {
+        return start(on: observingQueue, observer: command.execute)
     }
 }
