+++ SwiftyRedux/Sources/Core/Concurrency/Atomic.swift
