+++ SwiftyRedux/Sources/Core/Observing/ObservableProducer.swift
