+++ SwiftyRedux/Sources/Core/Observing/Observer.swift
